Add test cases for all sorts
Implemented a getSorted method in SortBase for easier testing.
Moved random seeding to a separate function which will be called only
once.
The TESTX macro creates a Sort object of the requested type, calls
getSorted with a vector of the given size and checks if the return
vector is sorted.

Note: MergeSortTest fails atm.

Signed-off-by: Thameera Senanayaka <thameera123@gmail.com>